requirements-dev.txt
-r requirements.txt
pytest==8.3.4
pytest-django==4.8.0
pytest-cov==6.0.0
pytest-html==4.1.1
pytest-xdist==3.6.1
factory-boy==3.3.1
faker==33.1.0
requests==2.32.3
selenium==4.27.1
beautifulsoup4==4.12.3


